Move GroupedCallReducer out of Result/
